yolov8s/requirements.txt
absl-py==1.4.0
asttokens==2.2.1
backcall==0.2.0
cachetools==5.3.0
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.10.0
Cython==0.29.34
-e git+https://github.com/samson-wang/cython_bbox.git@17a9209162d2107c1d2d17ea66fb2a3c7fa0c73e#egg=cython_bbox
dataclasses-json==0.5.7
decorator==5.1.1
executing==1.2.0
filelock==3.10.7
filterpy==1.4.5
flatbuffers==23.3.3
fonttools==4.39.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
h5py==3.8.0
humanfriendly==10.0
idna==2.10
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
lap @ file:///D:/bld/lap_1650201541914/work
lazy_loader==0.2
loguru==0.7.0
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
more-itertools==9.1.0
motmetrics==1.4.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
ninja==1.11.1
numpy==1.24.2
oauthlib==3.2.2
onemetric==0.1.2
onnx==1.13.1
onnx-simplifier==0.3.5
onnxoptimizer==0.3.13
onnxruntime==1.14.1
onnxruntime-gpu==1.14.1
opencv-python==4.7.0.72
packaging==23.0
pandas==2.0.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==2.4.7
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
roboflow==1.0.2
rsa==4.9
scikit-image==0.20.0
scipy==1.9.1
seaborn==0.12.2
sentry-sdk==1.19.1
six==1.16.0
stack-data==0.6.2
supervision==0.1.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
thop==0.1.1.post2209072238
tifffile==2023.4.12
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.32
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.3.3
wget==3.2
win32-setctime==1.1.0
wincertstore==0.2
xmltodict==0.13.0
-e git+https://github.com/ifzhang/ByteTrack.git@d1bf0191adff59bc8fcfeaa0b33d3d1642552a99#egg=yolox
zipp==3.15.0
